Refactor DeleteCaseCommandHandler to improve null case handling and logging

+++ Avera.Application/Cases/Delete/DeleteCaseCommandHandler.cs
@@ -2,6 +2,7 @@ using Avera.Application.Abstractions.Databases;
 using Avera.Application.Abstractions.Messaging;
 using Avera.Application.Abstractions.Storage;
 using Avera.Application.Delete;
+using Avera.Domain.Application.Cases;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Logging;
 using SharedKernel;
@@ -18,17 +19,24 @@ namespace Avera.Application.Cases.Delete
             var caseToDelete = await dbContext.Cases.FirstOrDefaultAsync(c => c.Id == command.CaseId, cancellationToken);
 
             logger.LogInformation("Attempting to delete case with ID: {CaseId}", command.CaseId);
-            // if (caseToDelete is null)
-            // {
-            //     return Task.FromResult(Result.Failure("Case not found."));
-            // }
+
+            if(caseToDelete is null)
+            {
+                logger.LogWarning("Case with ID: {CaseId} not found", command.CaseId);
+                return Result.Failure(CaseErrors.CaseNotFound);
+            }
 
             logger.LogInformation("Deleting blob storage for case with code: {CaseCode}", caseToDelete!.CaseCode);
+
             await blobStorageService.DeleteAsync(caseToDelete!.CaseCode, cancellationToken);
+
             logger.LogInformation("Deleted blob storage for case with code: {CaseCode}", caseToDelete.CaseCode);
 
+
             logger.LogInformation("Removing case with ID: {CaseId} from database", caseToDelete.Id);
+
             dbContext.Cases.Remove(caseToDelete!);
+
             logger.LogInformation("Removed case with ID: {CaseId} from database", caseToDelete.Id);
 
             await dbContext.SaveChangesAsync(cancellationToken);
